Added a rev counter
This turned out to need changes in Extplane-plugin:
the plugin didn't return arrays of float, and the
engine revs are such an array. As a result, a fair
bit of ugliness was required in extplaneclient to
add the data string for arrays, and then to process
it in engine_rpm. But it works!