Added bezels to the instruments
Added a bezel, created from a photo and edited in
Gimp. Then added code to read it into all the 
instruments and put it around them. It isn't 
elegant as it goes outside the 200x200 box defined
for instruments, which means it leaves droppings 
when moved on the panel display. But it looks 
better than no bezel, and works fine in normal 
operation.