Moved bezel inside instrument boundary
The bezel was outside the 200x200 limits of each
instrument, which led to detritus when redrawing.
Now all the instruments are a little smaller, and
the bezel is also smaller, to fit within the space.

A new, square bezel is also being used. It isn't
perfect, as the circular hole in the middle isn't
circular, and isn't centred exactly on the centre
of the instrument, but it isn't bad.

Also moved bezel reading code into constructor, so
it isn't called every paint.