Added coloured arcs to NeedleInstrument
The arcs are in. AirspeedIndicator also saves and
restores them. The engine rev counter doesn't, but
will do soon.

A lot of time wasted because I was maintaining the 
arc count seperately, and didn't stop doing that
once I started using a list for arcs. That led to
a nasty bug where the list was updated but the 
counter not. The bug is fixed, but the root cause
still needs to be sorted out.